refactor: change user ID assignment from uint to int
refactor: change ID type from uint to int64 in User struct

+++ service/auth_service/internal/repository/sql_user_repository.go
@@ -24,7 +24,7 @@ func (r *SqlUserRepository) CreateUser(ctx context.Context, exec SQLExecutor, us
 		return err
 	}
 
-	user.ID = uint(id)
+	user.ID = id
 	return nil
 }
